event page routing init
